requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
transformers>=4.21.0
huggingface-hub>=0.16.0
python-dotenv>=1.0.0